linked headings, some files (git slides), facelift
The main thing is that the new file, misc.js, houses some jQuery code to
add anchor links (intra-page links) to heading elements, h1 to h6. The
wrinkle here is that for our older Markdown posts, we auto-generate id
attributes that roughly match the text, which we can use to build our
link. But for newer posts written in OrgMode, the id attributes are not
generated at all, so we have to generate them manually.

We also do the usual site design changes. The git slides are also new.