css: add line numbers back to code blocks
Apparently at some point Hakyll moved to a different syntax highlighting
library, called "skylighting". We use that library to generate the
necessary CSS needed to go along with the HTML that the library
generates for syntax highlighting. It is this CSS sheet that creates the
line numbers, and are generated here so that the definitions never go
stale like it did before.

Also make sure that intra-document anchor links are not obstructed by
the sticky title div.

And make some preparations for the upcoming post about Bresenham's
circle algorithm (the stuff around tooltips and such for the canvas
element).